refactor (migrate): consolidates and reorganizes migration tasks
