+++ erpnext_mexico_compliance/migrate/__init__.py
@@ -4,3 +4,4 @@ from . import after_migrate
 def execute_after_migrate_tasks():
 	after_migrate.set_missing_uuids("Sales Invoice")
 	after_migrate.set_missing_uuids("Payment Entry")
+	after_migrate.enqueue_sat_catalogs_update()
