+++ erpnext_mexico_compliance/migrate/after_migrate.py
@@ -3,6 +3,7 @@ import typing as t
 import click
 import frappe
 
+from erpnext_mexico_compliance import sat
 from erpnext_mexico_compliance.utils.cfdi import get_uuid_from_xml
 
 
@@ -24,3 +25,15 @@ def set_missing_uuids(doctype: t.Literal["Sales Invoice", "Payment Entry"]):
 		with click.progressbar(query.run(), label=f"Setting {doctype} missing UUIDs") as bar:
 			for name, xml in bar:
 				frappe.db.set_value("Sales Invoice", name, "mx_uuid", get_uuid_from_xml(xml))
+
+
+def enqueue_sat_catalogs_update():
+	"""Queues a task to update the SAT Catalogs for the current site."""
+	frappe.enqueue(sat.update_tax_regimes)
+	frappe.enqueue(sat.update_cfdi_uses)
+	frappe.enqueue(sat.update_payment_options)
+	frappe.enqueue(sat.update_payment_methods)
+	frappe.enqueue(sat.update_product_or_service_keys, queue="long")
+	frappe.enqueue(sat.update_relationship_types)
+	frappe.enqueue(sat.update_units_of_measure, queue="long")
+	click.echo(f"Queued update of SAT Catalogs for {frappe.local.site}")
